LPS-50967 Moving the generic search classes to portal-service
