+++ portal-service/src/com/liferay/portal/kernel/search/generic/BooleanQueryFactoryImpl.java
@@ -12,31 +12,19 @@
  * details.
  */
 
-package com.liferay.portal.search.generic;
+package com.liferay.portal.kernel.search.generic;
 
-import com.liferay.portal.kernel.search.QueryTerm;
+import com.liferay.portal.kernel.search.BooleanQuery;
+import com.liferay.portal.kernel.search.BooleanQueryFactory;
 
 /**
  * @author Michael C. Han
  */
-public class QueryTermImpl implements QueryTerm {
-
-	public QueryTermImpl(String field, String value) {
-		_field = field;
-		_value = value;
-	}
+public class BooleanQueryFactoryImpl implements BooleanQueryFactory {
 
 	@Override
-	public String getField() {
-		return _field;
+	public BooleanQuery create() {
+		return new BooleanQueryImpl();
 	}
-
-	@Override
-	public String getValue() {
-		return _value;
-	}
-
-	private final String _field;
-	private final String _value;
 
 }
