+++ portal-service/src/com/liferay/portal/kernel/search/generic/TermQueryImpl.java
@@ -12,18 +12,18 @@
  * details.
  */
 
-package com.liferay.portal.search.generic;
+package com.liferay.portal.kernel.search.generic;
 
 import com.liferay.portal.kernel.search.BaseQueryImpl;
 import com.liferay.portal.kernel.search.QueryTerm;
-import com.liferay.portal.kernel.search.WildcardQuery;
+import com.liferay.portal.kernel.search.TermQuery;
 
 /**
  * @author Michael C. Han
  */
-public class WildcardQueryImpl extends BaseQueryImpl implements WildcardQuery {
+public class TermQueryImpl extends BaseQueryImpl implements TermQuery {
 
-	public WildcardQueryImpl(QueryTerm queryTerm) {
+	public TermQueryImpl(QueryTerm queryTerm) {
 		_queryTerm = queryTerm;
 	}
