+++ portal-service/src/com/liferay/portal/kernel/search/generic/WildcardQueryImpl.java
@@ -12,24 +12,31 @@
  * details.
  */
 
-package com.liferay.portal.search.generic;
+package com.liferay.portal.kernel.search.generic;
 
-import com.liferay.portal.kernel.search.TermQuery;
-import com.liferay.portal.kernel.search.TermQueryFactory;
+import com.liferay.portal.kernel.search.BaseQueryImpl;
+import com.liferay.portal.kernel.search.QueryTerm;
+import com.liferay.portal.kernel.search.WildcardQuery;
 
 /**
  * @author Michael C. Han
  */
-public class TermQueryFactoryImpl implements TermQueryFactory {
+public class WildcardQueryImpl extends BaseQueryImpl implements WildcardQuery {
 
-	@Override
-	public TermQuery create(String field, long value) {
-		return create(field, String.valueOf(value));
+	public WildcardQueryImpl(QueryTerm queryTerm) {
+		_queryTerm = queryTerm;
 	}
 
 	@Override
-	public TermQuery create(String field, String value) {
-		return new TermQueryImpl(new QueryTermImpl(field, value));
+	public QueryTerm getQueryTerm() {
+		return _queryTerm;
 	}
+
+	@Override
+	public Object getWrappedQuery() {
+		return this;
+	}
+
+	private final QueryTerm _queryTerm;
 
 }
